+++ poo/sudoku/src/ui/ConsoleInput.java
@@ -1,0 +1,18 @@
+package ui;
+
+import java.util.Scanner;
+
+public class ConsoleInput {
+
+    private final Scanner scanner;
+
+    public ConsoleInput() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    public int lerOpcao() {
+
+        return scanner.nextInt();
+
+    }
+}